[wip] - factor out partner classes
