+++ cellpack/autopack/loaders/recipe_loader.py
@@ -6,6 +6,7 @@ from json import encoder
 
 import cellpack.autopack as autopack
 from cellpack.autopack.interface_objects.ingredient_types import INGREDIENT_TYPE
+from cellpack.autopack.interface_objects.partners import Partners
 from cellpack.autopack.utils import deep_merge, expand_object_using_key
 from cellpack.autopack.interface_objects.representations import Representations
 from cellpack.autopack.interface_objects.default_values import default_recipe_values
@@ -167,6 +168,8 @@ class RecipeLoader(object):
                     atomic=reps.get("atomic", None),
                     packing=reps.get("packing", None),
                 )
+                partner_settings = obj["partners"] if "partners" in obj else {}
+                obj["partners"] = Partners(partner_settings)
             if not INGREDIENT_TYPE.is_member(obj["type"]):
                 raise TypeError(f"{obj['type']} is not an allowed type")
